requirements.txt
streamlit
langchain
langchain-community
langchain-groq
langchain-core
faiss-cpu
sentence-transformers
python-dotenv
requests